feat: completa CRUD de Todos com autenticacao e isolamento por usuario (owner_id)

+++ TodoApp/routers/todos.py
@@ -75,14 +75,14 @@ async def update_date(user: user_dependency, db: db_dependency, todo_request: To
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
     if user is None:
         return HTTPException(status_code=401, detail="Authentication Failed")
-        
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).flter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
